# SD in MSD: Speed up a bit