--- a/firmware/hw_layer/mass_storage/mass_storage_init.h
+++ b/firmware/hw_layer/mass_storage/mass_storage_init.h
@@ -4,6 +4,6 @@
 
 #if HAL_USE_USB_MSD
 void initUsbMsd();
-void attachMsdSdCard(BaseBlockDevice* blkdev);
+void attachMsdSdCard(BaseBlockDevice* blkdev, uint8_t *blkbuf, size_t blkbufsize);
 void deattachMsdSdCard(void);
 #endif