Read using UAEFI before this patch:
```
$ dd if=test.bin of=/dev/zero status=progress
133939712 bytes (134 MB, 128 MiB) copied, 269 s, 498 kB/s
262144+0 records in
262144+0 records out
134217728 bytes (134 MB, 128 MiB) copied, 269.492 s, 498 kB/s
```

Read using UAEFI with blkbuf1 = 2K bytes:
```
$ dd if=B5A4-271B/test.bin of=/dev/zero status=progress
133677568 bytes (134 MB, 127 MiB) copied, 202 s, 661 kB/s
262144+0 records in
262144+0 records out
134217728 bytes (134 MB, 128 MiB) copied, 202.69 s, 662 kB/s
```