@dron0gus 2025-02-14
Ups. Some boards runs out of RAM. We need to reuse same RAM for SD read/write buffer in MSD mode and for FatFS structures in ECU logging mode. As these functions are used exclusively.